+++ includes/class-oodsp-plugin.php
@@ -14,7 +14,7 @@
 
 /**
  *
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -110,9 +110,9 @@ class OODSP_Plugin {
 		require_once plugin_dir_path( __DIR__ ) . 'admin/settings/class-oodsp-settings.php';
 		require_once plugin_dir_path( __DIR__ ) . 'admin/class-oodsp-docspace.php';
 		require_once plugin_dir_path( __DIR__ ) . 'admin/class-oodsp-ajax.php';
-		require_once plugin_dir_path( __DIR__ ) . 'admin/class-oodsp-hook.php';
 		require_once plugin_dir_path( __DIR__ ) . 'includes/managers/class-oodsp-request-manager.php';
 		require_once plugin_dir_path( __DIR__ ) . 'includes/managers/class-oodsp-security-manager.php';
+		require_once plugin_dir_path( __DIR__ ) . 'includes/utils/class-oodsp-utils.php';
 		require_once plugin_dir_path( __DIR__ ) . 'includes/users/class-oodsp-users-list-table.php';
 		require_once plugin_dir_path( __DIR__ ) . 'includes/class-oodsp-i18n.php';
 		require_once plugin_dir_path( __DIR__ ) . 'includes/class-oodsp-loader.php';
@@ -169,9 +169,6 @@ class OODSP_Plugin {
 		$plugin_ajax = new OODSP_Ajax();
 		$this->loader->add_action( 'wp_ajax_oodsp_credentials', $plugin_ajax, 'oodsp_credentials' );
 		$this->loader->add_action( 'wp_ajax_nopriv_oodsp_credentials', $plugin_ajax, 'no_priv_oodsp_credentials' );
-
-		$plugin_hook = new OODSP_Hook();
-		$this->loader->add_action( 'post_updated', $plugin_hook, 'rooms_share' );
 	}
 
 	/**