+++ onlyoffice-docspace-wordpress-block/src/save.js
@@ -1,5 +1,5 @@
 /*
- * (c) Copyright Ascensio System SIA 2023
+ * (c) Copyright Ascensio System SIA 2024
  * 
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU General Public License
@@ -37,6 +37,18 @@ const Save = ({ attributes }) => {
 
     if ( attributes.hasOwnProperty('height') && attributes.height.length > 0 ) {
         parameters += 'height=' + attributes.height + ' ';
+    }
+
+    if ( attributes.hasOwnProperty('align') && attributes.align.length > 0 ) {
+        parameters += 'align=' + attributes.align + ' ';
+    }
+
+    if ( attributes.hasOwnProperty('theme') && attributes.theme.length > 0 ) {
+        parameters += 'theme=' + attributes.theme + ' ';
+    }
+
+    if ( attributes.hasOwnProperty('requestToken') && attributes.requestToken.length > 0 ) {
+        parameters += 'requestToken=' + attributes.requestToken + ' ';
     }
 
     return <RawHTML>{ `[onlyoffice-docspace ${ parameters } /]` }</RawHTML>;